+++ src/routes/+layout.js
@@ -1,0 +1,2 @@
+export const prerender = true;
+export const trailingSlash = "always";
