+++ src/routes/+page.server.js
@@ -1,0 +1,3 @@
+export async function load() {
+	return {};
+}
