+++ vite.config.ts
@@ -1,8 +1,21 @@
 import tailwindcss from '@tailwindcss/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
+import path from "path";
 import dsv from '@rollup/plugin-dsv';
 
 export default defineConfig({
-	plugins: [sveltekit(), tailwindcss(), dsv()]
+	plugins: [sveltekit(), tailwindcss(), dsv()],
+	resolve: {
+		alias: {
+			$actions: path.resolve("./src/actions"),
+			$components: path.resolve("./src/components"),
+			$data: path.resolve("./src/data"),
+			$routes: path.resolve("./src/routes"),
+			$runes: path.resolve("./src/runes"),
+			$styles: path.resolve("./src/styles"),
+			$svg: path.resolve("./src/svg"),
+			$utils: path.resolve("./src/utils")
+		}
+	}
 });
